+++ src/lib/components/display/timestamp/TimeStamp.svelte
@@ -1,6 +1,4 @@
 <script lang="ts">
-    import {Temporal} from "../../../vendor/js-temporal-polyfill";
-
     import type {IGlobalProperties} from "../../../types/global";
     import type {IHTML5Events, IHTML5Properties} from "../../../types/html5";
     import type {IMarginProperties} from "../../../types/spacings";
@@ -9,6 +7,7 @@
     import type {IHourFormatOptions} from "../../../util/datetime/hours";
     import type {IMinuteFormatOptions} from "../../../util/datetime/minutes";
     import type {ISecondFormatOptions} from "../../../util/datetime/seconds";
+    import {from_timestamp} from "../../../util/datetime/timestamps";
     import {defaultopt} from "../../../util/functional";
     import {DEFAULT_FORMAT_TIME, DEFAULT_LOCALE} from "../../../util/locale";
 
@@ -49,7 +48,7 @@
     export let timestamp: $$Props["timestamp"];
 
     $: _options = defaultopt({hour, hour_12, minute, second}, DEFAULT_FORMAT_TIME);
-    $: _time = Temporal.PlainTime.from(timestamp);
+    $: _time = from_timestamp(timestamp);
 </script>
 
 <time