+++ src/lib/util/datetime/datestamps.ts
@@ -4,18 +4,17 @@ export function from_datestamp(
     timestamp: string | Temporal.PlainDateLike,
     modify?: Temporal.PlainDateLike
 ): Temporal.PlainDate {
-    if (modify) return Temporal.PlainDate.from(timestamp).withCalendar("iso8601").with(modify);
-    return Temporal.PlainDate.from(timestamp).withCalendar("iso8601");
+    return modify
+        ? Temporal.PlainDate.from(timestamp).withCalendar("iso8601").with(modify)
+        : Temporal.PlainDate.from(timestamp).withCalendar("iso8601");
 }
 
 export function to_datestamp(date: Temporal.PlainDate, modify?: Temporal.PlainDateLike): string {
-    if (modify) {
-        return date.withCalendar("iso8601").with(modify).toString({
-            calendarName: "never",
-        });
-    }
-
-    return date.withCalendar("iso8601").toString({
-        calendarName: "never",
-    });
+    return modify
+        ? date.withCalendar("iso8601").with(modify).toString({
+              calendarName: "never",
+          })
+        : date.withCalendar("iso8601").toString({
+              calendarName: "never",
+          });
 }