+++ src/lib/util/datetime/timezones.ts
@@ -1,5 +1,0 @@
-const EXPRESSION_TIMEZONE = /\[[\w/]+\]$/;
-
-export function has_timezone(timestamp: string): boolean {
-    return EXPRESSION_TIMEZONE.test(timestamp);
-}